refactor(pdf): isolate figure candidates
